requirements.txt
scikit-learn==1.7.0
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
PyGithub==2.2.0
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
click==8.1.7
pydantic==2.5.2
sqlalchemy==2.0.23
python-dateutil==2.8.2 